+++ im_flutter_sdk_android/android/src/main/java/com/easemob/im_flutter_sdk/MethodKey.java
@@ -21,14 +21,6 @@ public class MethodKey {
 
     static final String onConnected = "onConnected";
     static final String onDisconnected = "onDisconnected";
-    static final String onUserDidLoginFromOtherDevice = "onUserDidLoginFromOtherDevice";
-    static final String onUserDidRemoveFromServer = "onUserDidRemoveFromServer";
-    static final String onUserDidForbidByServer = "onUserDidForbidByServer";
-    static final String onUserDidChangePassword = "onUserDidChangePassword";
-    static final String onUserDidLoginTooManyDevice = "onUserDidLoginTooManyDevice";
-    static final String onUserKickedByOtherDevice = "onUserKickedByOtherDevice";
-    static final String onUserAuthenticationFailed = "onUserAuthenticationFailed";
-    static final String onAppActiveNumberReachLimit = "onAppActiveNumberReachLimit";
     static final String onMultiDeviceGroupEvent = "onMultiDeviceGroupEvent";
     static final String onMultiDeviceContactEvent = "onMultiDeviceContactEvent";
     static final String onMultiDeviceThreadEvent = "onMultiDeviceThreadEvent";